requirements.txt
requests>=2.28.0
PyYAML>=6.0
pytz>=2022.1
urllib3>=1.26.0
